refactor/tcp: pull write queue to a separate module
Will make it easier to reuse it.
